asdasd

+++ app/Http/Controllers/UsersController.php
@@ -136,7 +136,90 @@ class UsersController extends Controller
      */
     public function destroy(User $user)
     {
-        $user->delete();
+        $check = DB::Table('users')
+            ->join('employees', 'employees.user_id', '=', 'users.id')
+            ->join('payrolls', 'payrolls.employee_id', '=', 'employees.id')
+            ->where('user_id', '=', $user->id)
+            ->select('payrolls.id')
+            ->exists();
+
+        $check1 = Db::Table('employees')->where('user_id', '=', $user->id)->exists();
+        $check2 = Db::Table('costs_imputs')->where('user_id', '=', $user->id)->exists();
+        $check3 = Db::Table('others_documents')->where('user_id', '=', $user->id)->exists();
+
+        if ($check) {
+            $payrollFilename = DB::Table('users')
+                ->join('employees', 'employees.user_id', '=', 'users.id')
+                ->join('payrolls', 'payrolls.employee_id', '=', 'employees.id')
+                ->where('user_id', '=', $user->id)
+                ->value('payrolls.filename');
+            foreach ($payrollFilename as $index) {
+                unlink($index);
+            }
+            $payrollsId = DB::Table('users')
+                ->join('employees', 'employees.user_id', '=', 'users.id')
+                ->join('payrolls', 'payrolls.employee_id', '=', 'employees.id')
+                ->where('user_id', '=', $user->id)
+                ->value('payrolls.id');
+
+            Db::Table('payrolls')->whereIn('id', '=', $payrollsId)->delete();
+
+
+            $check = Db::Table('employees')->where('user_id', '=', $user->id)->exists();
+
+            if ($check) {
+                Db::Table('employees')->where('user_id', '=', $user->id)->delete();
+            }
+
+            $check = Db::Table('costs_imputs')->where('user_id', '=', $user->id)->exists();
+
+            if ($check) {
+                $costsimputsId = Db::Table('costs_imputs')->where('user_id', '=', $user->id)->value('filename');
+                foreach ($costsimputsId as $index) {
+                    unlink($index);
+                }
+                Db::Table('costs_imputs')->where('user_id', '=', $user->id)->delete();
+            }
+
+            $check = Db::Table('others_documents')->where('user_id', '=', $user->id)->exists();
+
+            if ($check) {
+                $othersdocumentsId = Db::Table('others_documents')->where('user_id', '=', $user->id)->value('filename');
+                foreach ($othersdocumentsId as $index) {
+                    unlink($index);
+                }
+                Db::Table('others_documents')->where('user_id', '=', $user->id)->delete();
+            }
+
+            $user->delete();
+        } elseif ($check1 || $check2 || $check3) {
+
+            if ($check1) {
+                Db::Table('employees')->where('user_id', '=', $user->id)->delete();
+            }
+
+
+            if ($check2) {
+                $costsimputsId = Db::Table('costs_imputs')->where('user_id', '=', $user->id)->value('filename');
+                foreach ($costsimputsId as $index) {
+                    unlink($index);
+                }
+                Db::Table('costs_imputs')->where('user_id', '=', $user->id)->delete();
+            }
+
+
+            if ($check3) {
+                $othersdocumentsId = Db::Table('others_documents')->where('user_id', '=', $user->id)->value('filename');
+                foreach ($othersdocumentsId as $index) {
+                    unlink($index);
+                }
+                Db::Table('others_documents')->where('user_id', '=', $user->id)->delete();
+            }
+
+            $user->delete();
+        } else {
+            $user->delete();
+        }
 
         return redirect()->route('users.index')
             ->withSuccess(__('Empresas eliminado correctamente.'));
